Enhance file upload handling
- Introduced a new function `createSignedUrlWithRetry` to handle JWT token issues during signed URL generation with retry logic.
- Updated the `POST` request handler in `regenerateUrl` to utilize the new signed URL function, improving error handling.
- Increased the maximum file size limit for uploads from 10MB to 50MB in the `upload` route.
- Refactored module form imports for consistency and updated UI components for better user experience.
- Removed the deprecated `ModuleForm` component and adjusted related imports accordingly.
- Enhanced resource type detection in the `ResourceUploadDialog` using shared utility functions for improved accuracy.
