+++ app/api/resources/upload/route.ts
@@ -73,10 +73,10 @@ export async function POST(request: NextRequest) {
     const fileSize = file.size;
 
     // Validate file size (10MB limit)
-    const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB
+    const MAX_FILE_SIZE = 50 * 1024 * 1024; // 10MB
     if (fileSize > MAX_FILE_SIZE) {
       return NextResponse.json(
-        { error: "File size exceeds limit (10MB)" },
+        { error: "File size exceeds limit (50MB)" },
         { status: 400 }
       );
     }
@@ -131,7 +131,7 @@ export async function POST(request: NextRequest) {
     const resource = await prisma.resource.create({
       data: {
         title,
-        type,
+        type: fileType,
         fileUrl: normalizedFileUrl,
         moduleId,
         userId,
