+++ components/ResourceUploadButton.tsx
@@ -45,7 +45,7 @@ export function ResourceUploadButton({
         {children || (
           <>
             <Upload className="mr-2 h-4 w-4" />
-            Upload Resource
+            Upload
           </>
         )}
       </Button>
