+++ components/dialogs/DeleteConfirmationDialog.tsx
@@ -1,0 +1,55 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from "@/components/ui/alert-dialog";
+
+interface DeleteConfirmationDialogProps {
+  isOpen: boolean;
+  setIsOpen: (isOpen: boolean) => void;
+  resourceTitle: string;
+  onDelete: () => void;
+}
+
+/**
+ * A dialog for confirming resource deletion.
+ *
+ * @param isOpen - Whether the dialog is currently open
+ * @param setIsOpen - Function to set the open state
+ * @param resourceTitle - The title of the resource being deleted
+ * @param onDelete - Function to call when deletion is confirmed
+ */
+export function DeleteConfirmationDialog({
+  isOpen,
+  setIsOpen,
+  resourceTitle,
+  onDelete,
+}: DeleteConfirmationDialogProps) {
+  return (
+    <AlertDialog open={isOpen} onOpenChange={setIsOpen}>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This will permanently delete &quot;{resourceTitle}&quot; and cannot
+            be undone.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={onDelete}
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+          >
+            Delete
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
